+++ src/TwigComponent/tests/Fixture/Kernel.php
@@ -20,6 +20,7 @@ use Symfony\Component\HttpKernel\Kernel as BaseKernel;
 use Symfony\UX\TwigComponent\Tests\Fixture\Component\ComponentA;
 use Symfony\UX\TwigComponent\Tests\Fixture\Component\ComponentB;
 use Symfony\UX\TwigComponent\Tests\Fixture\Component\ComponentC;
+use Symfony\UX\TwigComponent\Tests\Fixture\Component\WithAttributes;
 use Symfony\UX\TwigComponent\Tests\Fixture\Service\ServiceA;
 use Symfony\UX\TwigComponent\TwigComponentBundle;
 
@@ -53,6 +54,7 @@ final class Kernel extends BaseKernel
         $c->register(ComponentA::class)->setAutoconfigured(true)->setAutowired(true);
         $c->register('component_b', ComponentB::class)->setAutoconfigured(true)->setAutowired(true);
         $c->register(ComponentC::class)->setAutoconfigured(true)->setAutowired(true);
+        $c->register(WithAttributes::class)->setAutoconfigured(true)->setAutowired(true);
         $c->register('component_d', ComponentB::class)->addTag('twig.component', [
             'key' => 'component_d',
             'template' => 'components/custom2.html.twig',